LLM-and-Recommendation-Sys/requirements.txt
# Core Python Libraries
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0

# Web Framework (FastAPI)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
httpx>=0.25.0
certifi>=2023.0.0

# Database & Vector Operations
psycopg2-binary>=2.9.9
pgvector>=0.2.3

# AI/ML & NLP Libraries
langchain-core>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.15
langchain-openai>=0.0.3
openai>=1.5.0
sentence-transformers>=2.2.2

# Optional ML Libraries (uncomment if needed)
# scikit-learn>=1.3.0
# torch>=2.0.0
# tensorflow>=2.13.0

# Development & Testing (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0